LPS-131544 Merge RoleFinderTest into RoleLocalServiceTest.
